api/requirements.txt
fastapi
uvicorn[standard]
qiskit
qiskit-aer
pydantic
python-dotenv
